requirements.txt
discord.py
pytesseract
Pillow
selenium
webdriver-manager